+++ game/cli.py
@@ -1,4 +1,6 @@
-from chess import Chess
+# cli.py
+from game.chess import Chess
+from game.exceptions import InvalidMove, InvalidTurn, EmptyPosition
 
 def main():
     chess = Chess()
@@ -8,20 +10,23 @@ def main():
 def play(chess):
     try:
         # print(chess.show_board())
-        print("turn: ", chess.turn)
+        print("Turn:", chess.current_turn)
         from_row = int(input("From row: "))
         from_col = int(input("From col: "))
-        to_row = int(input("To Row: "))
-        to_col = int(input("To Col: "))
-        # :)
-        chess.move(
-            from_row,
-            from_col,
-            to_row,
-            to_col,
-        )
+        to_row = int(input("To row: "))
+        to_col = int(input("To col: "))
+        promotion_choice = input("Promotion choice (queen/rook/bishop/knight, leave empty if not applicable): ").strip().lower()
+        chess.move(from_row, from_col, to_row, to_col, promotion_choice if promotion_choice else None)
+    except EmptyPosition:
+        print("Error: The starting position is empty.")
+    except InvalidTurn:
+        print("Error: It's not your turn to move this piece.")
+    except InvalidMove:
+        print("Error: Invalid move for this piece.")
+    except ValueError as ve:
+        print(f"Error: {ve}")
     except Exception as e:
-        print("error", e)
+        print("Error:", e)
 
 if __name__ == '__main__':
     main()